refactor: product_ranking package, product package 분리
