+++ src/main/java/kr/hhplus/be/commerce/application/product/ProductReadAllDailyTopSellingQueryManager.java
@@ -9,10 +9,10 @@ import java.util.Map;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import kr.hhplus.be.commerce.domain.product.model.ProductRanking;
 import kr.hhplus.be.commerce.domain.product.model.ProductSummaryView;
 import kr.hhplus.be.commerce.domain.product.repository.ProductReader;
-import kr.hhplus.be.commerce.domain.product.store.ProductRankingStore;
+import kr.hhplus.be.commerce.domain.product_ranking.model.ProductRankingView;
+import kr.hhplus.be.commerce.domain.product_ranking.store.ProductRankingStore;
 import kr.hhplus.be.commerce.presentation.api.product.response.ProductSummaryResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -26,15 +26,17 @@ public class ProductReadAllDailyTopSellingQueryManager {
 
 	@Transactional(readOnly = true)
 	public List<ProductSummaryResponse> read(Query query) {
+
+		// 레디스에서 문제가 발생한 경우. 데이터가 불일치 하거나,
 		List<Long> productIds = productRankingStore.readAllByRankingDate(query.today, query.limit)
 			.stream()
-			.map(ProductRanking::productId)
+			.map(ProductRankingView::productId)
 			.toList();
 
 		Map<Long, ProductSummaryView> productIdToProductMap = productReader.readAllByIdIn(productIds)
 			.stream()
 			.collect(toMap(ProductSummaryView::id, it -> it));
-		
+
 		return productIds
 			.stream()
 			.filter(productIdToProductMap::containsKey)
